requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
openai
python-dotenv
numpy